12378: выделил новый класс SyscallStats для утилиты трассировки, добавил --follow-forks в strace, добавил возможность трассировать конкретный процесс
